hopr_crypto_types

Module crypto_traits

Re-exports from the generic cryptographic traits.

Traits

BlockSizeUser

Types which process data in blocks.

Digest

Convenience wrapper trait covering functionality of cryptographic hash functions with fixed output size.

FixedOutput

Trait for hash functions with fixed-size output.

FixedOutputReset

Trait for hash functions with fixed-size output able to reset themselves.

IvSizeUser

Types which use initialization vector (nonce) for initialization.

KeyInit

Types which can be initialized from key.

KeyIvInit

Types which can be initialized from key and initialization vector (nonce).

KeySizeUser

Types which use key for initialization.

OutputSizeUser

Types which return data with the given size.

PRP

Pseudo-random permutation (PRP)

Randomizable

Trait for types that can be randomly generated.

StreamCipher

Synchronous stream cipher core trait.

StreamCipherSeek

Trait for seekable stream ciphers.

UniversalHash

The UniversalHash trait defines a generic interface for universal hash functions.

Update

Types which consume data with byte granularity.

Type Aliases

Block

Block on which BlockSizeUser implementors operate.

Iv

Initialization vector (nonce) used by IvSizeUser implementors.

Key

Key used by KeySizeUser implementors.

Output

Output array of OutputSizeUser implementors.