Module hopr_crypto_types::keypairs

Implements ChainKeypair and OffchainKeypair, the important representations of chain key and packet key.

Structs

ChainKeypair

Represents a keypair consisting of a secp256k1 private and public key

OffchainKeypair

Represents a keypair consisting of an Ed25519 private and public key

Traits

Keypair

Represents a generic key pair The keypair contains a private key and public key. Must be comparable in constant time and zeroized on drop.