hopr_crypto_types::lioness

Struct Lioness

pub struct Lioness<H: KeySizeUser + OutputSizeUser, S: KeySizeUser + IvSizeUser, B>
where
    H::OutputSize: IsEqual<S::KeySize, Output = B1>,
    H::KeySize: IsEqual<S::KeySize, Output = B1>,
    B: IsGreater<<S as KeySizeUser>::KeySize, Output = B1> + ArrayLength<u8>,
{ /* private fields */ }

Implementation of Lioness wide-block cipher over a keyed Digest and a StreamCipher.

Requirements

The output size of the Digest H must match the key size of the StreamCipher.

The key size of the keyed digest H must be equal to the key size of the StreamCipher.

The block size B can be arbitrary but must be strictly greater than the key size of the StreamCipher.

However, for cryptographic security, B must be at least twice the key size of the StreamCipher.

The key size of the Lioness cipher is 4-times the size of StreamCipher’s key.

The IV size of the Lioness cipher is 2-times the size of the StreamCipher’s IV size.

Implementations

impl<H: Digest + KeyInit, S: StreamCipher + KeyIvInit, B> Lioness<H, S, B>
where H::OutputSize: IsEqual<<S as KeySizeUser>::KeySize, Output = B1> + Mul<U4>, H::KeySize: IsEqual<S::KeySize, Output = B1>, B: IsGreater<<S as KeySizeUser>::KeySize, Output = B1> + Sub<<S as KeySizeUser>::KeySize> + ArrayLength<u8>, <B as Sub<<S as KeySizeUser>::KeySize>>::Output: ArrayLength<u8>, <S as IvSizeUser>::IvSize: Mul<U2>, <H::OutputSize as Mul<U4>>::Output: ArrayLength<u8>, <<S as IvSizeUser>::IvSize as Mul<U2>>::Output: ArrayLength<u8>,

pub fn encrypt_block(&self, block: InOut<'_, '_, Block<Self>>)

Performs encryption of the given block.

pub fn decrypt_block(&self, block: InOut<'_, '_, Block<Self>>)

Performs decryption of the given block.