hopr_crypto_types::types

Struct CurvePoint

pub struct CurvePoint { /* private fields */ }

Represents an elliptic curve point on the secp256k1 curve It stores the compressed (and optionally also the uncompressed) form.

let a: [u8; 32] = hex!("876027a13900aad908842c3f79307cc8e96de5c3331090e91a24c315f2a8d43a");
let b: [u8; 32] = hex!("561d3fe2990e6a768b90f7d510b69e967e0922a3b61e8141398113aede8e1d3e");

let A = CurvePoint::from_exponent(&a).unwrap();
let B = CurvePoint::from_exponent(&b).unwrap();

// A_plus_B = a * G + b * G
let A_plus_B = CurvePoint::combine(&[&A, &B]);

let scalar_a: Scalar = *NonZeroScalar::<Secp256k1>::try_from(&a[..]).unwrap();
let scalar_b: Scalar = *NonZeroScalar::<Secp256k1>::try_from(&b[..]).unwrap();

// a_plus_b = (a + b) * G
let a_plus_b = CurvePoint::from_exponent(&(scalar_a + scalar_b).to_bytes()).unwrap();

// group homomorphism
// (a + b) * G = a * G + b * G
assert_eq!(A_plus_B, a_plus_b);

Implementations

impl CurvePoint

pub const SIZE_COMPRESSED: usize = 33usize

Size of the point if serialized via CurvePoint::as_compressed.

pub const SIZE_UNCOMPRESSED: usize = 65usize

Size of the point if serialized via CurvePoint::as_uncompressed.

pub fn to_address(&self) -> Address

Converts the uncompressed representation of the curve point to Ethereum address.

pub fn from_exponent(exponent: &[u8]) -> Result<Self>

Creates a curve point from a non-zero scalar. The given exponent must represent a non-zero scalar and must result into a secp256k1 identity point.

pub fn into_projective_point(self) -> ProjectivePoint<Secp256k1>

Converts the curve point to a representation suitable for calculations.

pub fn as_compressed(&self) -> &EncodedPoint<Secp256k1>

Converts the curve point into a compressed form. This is a cheap operation.

pub fn as_uncompressed(&self) -> &EncodedPoint<Secp256k1>

Converts the curve point into an uncompressed form. This is many cases an expensive operation.

pub fn combine(summands: &[&CurvePoint]) -> CurvePoint

Sums all given curve points together, creating a new curve point.