hopr_crypto_types

Module vrf

Contains implementation of Verifiable Random Function used in tickets

Structs

VrfParameters

Bundles values given to the smart contract to prove that a ticket is a win.

Functions

derive_vrf_parameters

Takes a private key, the corresponding Ethereum address and a payload and creates all parameters that are required by the smart contract to prove that a ticket is a win.