Crate hopr_crypto_types

This Rust crate contains implementation of common cryptographic types.

Modules

errors

Contains error enum implementation used across other hopr-crypto-... crates

keypairs

Implements ChainKeypair and OffchainKeypair, the important representations of chain key and packet key.

primitives

Implements low-level cryptographic primitives, such as SimpleStreamCipher, SimpleDigest and SimpleMac.

seal

Enables randomized encryption (sealing) and decryption of data using OffchainKeypair.

types

Implements basic cryptography-related types based on primitives, such as Hash, PublicKey and Signature.

utils

Contains small utility functions used in the other hopr-crypto-... crates

vrf

Contains implementation of Verifiable Random Function used in tickets