Crate hopr_crypto_types

This Rust crate contains implementation of common cryptographic types.

Modules

crypto_traits

Re-exports from the generic cryptographic traits.

errors

Contains error enum implementation used across other hopr-crypto-... crates

keypairs

Implements ChainKeypair and OffchainKeypair, the important representations of chain key and packet key.

lioness

Implementation of the Lioness PRP This module implements a generic Lioness wide-block cipher.

primitives

Re-exports of low-level cryptographic primitives.

seal

Enables randomized encryption (sealing) and decryption of data using OffchainKeypair.

signing

Separate module for signature algorithms.

types

Implements basic cryptography-related types based on primitives, such as Hash, PublicKey and Signature.

utils

Contains small utility functions used in the other hopr-crypto-... crates

vrf

Contains implementation of Verifiable Random Function used in tickets