hopr_crypto_types

Module primitives

Re-exports of low-level cryptographic primitives.

Structs

Blake3

An incremental hash state that can accept any number of writes.

Blake3Output

An incremental reader for extended output, returned by Hasher::finalize_xof.

IvKey

Convenience container for IV and key of a given primitive T.

Poly1305

The Poly1305 universal hash function.

Functions

blake3_hash

The default hash function.

Type Aliases

Aes128Ctr

AES with 128-bit key in counter-mode (with big-endian counter).

ChaCha20

ChaCha20 stream cipher (RFC 8439 version with 96-bit nonce)

Keccak256

Keccak-256 hasher state.

SecretKey

Represents a 256-bit secret key of fixed length. The value is auto-zeroized on drop.

SecretKey16

Represents a 128-bit secret key of fixed length. The value is auto-zeroized on drop.

Sha3_256

SHA-3-256 hasher state.