hopr_crypto_types

Module types

Implements basic cryptography-related types based on primitives, such as Hash, PublicKey and Signature.

Structs

Challenge

Contains the complete Proof-of-Relay challenge is a secp256k1 curve point.

HalfKey

Represents a half-key used for the Proof-of-Relay.

HalfKeyChallenge

Represents a challenge for the half-key in Proof of Relay.

HashBase

Represents a generic 256-bit hash value.

OffchainPublicKey

Represents an Ed25519 public key.

PublicKey

Represents a secp256k1 public key.

Response

Contains a response upon ticket acknowledgement It is equivalent to a non-zero secret scalar on secp256k1 (EC private key).

SimplePseudonym

Represents a simple UUID-like pseudonym consisting of 10 bytes.

Constants

PACKET_TAG_LENGTH

Length of a packet tag

Traits

Pseudonym

Pseudonym used to identify the creator of a SURB. This allows indexing SURB and LocalSURBEntry at both parties.

Type Aliases

Hash

Represents an Ethereum 256-bit hash value.

HashFast

Represents an alternative 256-bit hash value computed via a faster hashing algorithm.

PacketTag

Represents a fixed size packet verification tag